[core] introduce Unit Class, Bear and Hunter extend Unit
the Unit class will father both Bear and Hunter, as to keep their common
props and methods in one place, as well as abstract any functionality
pertaining to "units".

commit also introduces the 'Hunter' class (child class of Unit), which
similar to the Bear can get placed on a level.  as of now this is
supposed to be a programmatically controlled unit, maybe something to
chase the Bear through the cave.