[events] introduce criteriaCheck prop function for events
this is an optional setting for events, that is a function that is
executed just before the event is executed and if the function resolves
as false then the event execution does NOT run.

will be useful to have events being executed only when they make sense
to be so, e.g. some events should only execute when a certain unit is in
the room.