[core] rm unused/unwanted/never-implemented/draft-ideas code pieces
I'm cleaning up the codebase here as to get to a more 'final' state
of this repo, so that eventually can show it to others and not be too
messy.

overall cleanup based on error messages/warnings from the TS LUA server
which indicates unused code and reundant code.

some of these were future implementation ideas that I'll not be
implementing for the perceivable future.