[core] introduce speed setting for hunter unit when chasing
that is, the 'moveHunterAcrossMaze' func which unleashes the hunter
to traverse the maze, can now also dictate the speed at which
the hunter moves.  this then means a given level could call
this function with a higher speed and have the hunter move
faster thus making it more challenging for the player.

for now, default is speed 1 (i.e. hunter moves every 1000ms).
speed 2 would make hunter move every 500ms.
speed 3 would make hunter move every 333ms.
and so on... formula is = 1000ms / speed