[core] rm unused events and feature
feature being the 'event area' and associated clear event area.

the 'event area' (HTML box where events could register messages/logs) is
a nice idea in principle, but as it stands it's just getting in the
way of the game itself, so I'm opting to get rid of it for now.

also removed other unused events like dialogue and prompt.