[core] refactor & improve collision check
the previous check was only done when the hunter moved, so it was
missing collision when the bear unit moved on top of the hunter unit.

this commit also introduces the notion/feature of eventQueue
'subscribe/notify', where an option can be passed to the queue instance
as to add subscribers for it.  the subscribers are then notified when an
event gets processed in that queue, giving the subscribers the ability
to perform actions once the event is processed.
by subscribing to the 'moveEvent' EventQueues, we can then check for
collisions when the happen from either due to Bear or Hunter movement.